Added support for the CPE applications database. As both cpe-os and cpe-app are made from the same source file (the official CPE dictionary) this task required to refactor the DB updating mechanisms.
CPE databases use a cached dictionary, which is downloaded the first time we
need it and stored as a temporary file. The content of this file is still
processed once per CPE database (2 times).

Both the OS and the Applications databases share the same XML parsers to process
the original dictionary.